+++ packages/hero_ui/lib/src/components/components.dart
@@ -17,6 +17,7 @@ export 'color_slider/color_slider.dart';
 export 'color_swatch/color_swatch.dart';
 export 'color_swatch_picker/color_swatch_picker.dart';
 export 'description/description.dart';
+export 'empty_state/empty_state.dart';
 export 'field_error/field_error.dart';
 export 'form/form.dart';
 export 'input/input.dart';